SPATAData source data.frame

Access the source data.frame which contains information about various spatial transcriptomic experiments .

sourceDataFrame()

Value

Data.frame in which each row corresponds to a spatial data set stored in a SPATA2 object.

sample_name: Character. Name of the sample and its unique identifier.

comment: Character. Additional comments about the sample.

donor_id: Character. Unique identifier for the donor.

donor_species: Character. Species of the donor.

grade: Character. Grade of the sample.

grade_sub: Character. Sub-grade of the sample.

histo_class: Character. Histological classification.

histo_class_sub: Character. Sub-classification of the histological class.

institution: Character. Institution where the sample was collected.

lm_source: Date-time. Last instance when the row of the source data.frame was modified.

organ: Character. Organ from which the sample was taken.

organ_part: Character. Specific part of the organ from which the sample was taken.

pathology: Character. Pathological state of the sample.

platform: Character. Platform used for the experiment.

pub_citation: Character. Citation for the publication related to the sample.

pub_doi: Character. DOI of the publication related to the sample.

pub_journal: Character. Journal where the related publication was published.

pub_year: Numeric. Year of publication.

sex: Character. Sex of the donor.

side: Character. Side of the organ from which the sample was taken.

tags: Character. Tags related to the sample.

tissue_age: Numeric. Age of the tissue in years.

web_link: Character. Link to an object or additional data related to the sample for downloads.

workgroup: Character. Workgroup or team responsible for the sample.

mean_counts: Numeric. Mean counts of the measurements.

median_counts: Numeric. Median counts of the measurements.

modality_gene: Logical. Indicates if the modality includes genes.

modality_metabolite: Logical. Indicates if the modality includes metabolites.

modality_protein: Logical. Indicates if the modality includes proteins.

n_obs: Numeric. Number of observations.

n_tissue_sections: Numeric. Number of tissue sections.

obs_unit: Character. Unit of observation.

Examples

library(SPATA2)
library(SPATAData)
library(dplyr)
library(stringr)

sdf <- sourceDataFrame()

#----- example dplyr logic to filter the source data.frame

# 1. obtain glioblastoma samples from the temporal lobe

sdf_gbm <- filter(sdf, histo_class == "Glioblastoma" & organ_part == "temporal")

gbm_samples <- sdf_gbm$sample_name

print(gbm_samples)

# downlaod as collection
downloadSpataObjects(sample_names = gbm_samples, folder = "spata_objects/gbm") 

# 2. obtain data from specific publications

sdf_kuppe <- 
 filter(sdf, str_detect(pub_citation, pattern = "^Kuppe") & pub_journal == "Nature")
 
kuppe_samples <- sdf_kuppe$sample_name

print(kuppe_samples)